PESBot/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
multidict==6.0.5
oauth2client==4.1.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDrive==1.3.1
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
